django/requirements.txt
amqp==5.0.9
arrow==1.2.2
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
binaryornot==0.4.4
celery==5.2.3
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cookiecutter==1.7.3
Deprecated==1.2.13
Django==4.0.2
django-browser-reload==1.3.0
django-debug-toolbar==3.2.4
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-filter==21.1
django-nine==0.2.5
django-polymorphic==3.1.0
django-prometheus==2.2.0
django-tailwind==3.1.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==5.0.0
elasticsearch==7.17.0
elasticsearch-dsl==7.4.0
idna==3.3
inflection==0.5.1
Jinja2==3.0.3
jinja2-time==0.2.0
JSON-log-formatter==0.5.1
jsonschema==4.4.0
kombu==5.2.3
MarkupSafe==2.1.0
openapi==1.1.0
packaging==21.3
poyo==0.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
PyYAML==6.0
redis==4.1.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
